[#1,push,wip][l]: start of an effort to implement a push command using heroku style.
This is was very incomplete but pushing now as may not continue. I am half-way through porting incrementally from heroku codebase. Start with plugins code to load plugins as plan was next to implement the push command in node and load as plugin. Had not yet got plugins code working with the cobra setup (was about first 25% of way).

* Heroku: started with plugins and then copying other files plugins needs to work. Plugins refactor to work with cobra commands vs heroku's home grown stuff was in progress but a way from complete.

  cmd/plugins.go
  cmd/plugins_test.go  # just copied - no new content (copied from version test)
  cmd/context.go       # ported from
  filesystem.go

* Cobra stuff
  * cmd/push.go stubbed cobra function (not yet wired up ...)